+++ wizard/validators/__init__.py
@@ -1,0 +1,5 @@
+"""Validator registry exports for wizard components."""
+
+from wizard.validators.registry import PROFILE_VALIDATED_FIELDS, REQUIRED_FIELD_VALIDATORS
+
+__all__ = ["PROFILE_VALIDATED_FIELDS", "REQUIRED_FIELD_VALIDATORS"]